Create launcher.bat
This 'launcher' batch script is the first file used to download the initial 'trigger' file that does all the other downloading of scripts. This file we want to be simple and only used to download the MAIN script logic file, we call 'trigger'. You can also create a "BAT to EXE" file. We configured this batch file to delete itself once it runs.