Create trigger.ps1
This is the 'trigger' powershell script that downloads and does all the clean-up from your scripts. There is a loop under the sendmail function that works ok but, I'm sure it can be better. It will wait for all the scripts to finish then it will attach the log files and delete all remaining scripts and logs from the folder download location which we have set as the %USERPROFILE%\AppData\Local\Temp folder.